pyproject.toml
[project]
name = "cozyreq"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-ai-slim[mistral]>=1.39.0",
    "pyyaml>=6.0.3",
    "textual>=6.11.0",
    "typer>=0.20.1",
]

[project.scripts]
cozyreq = "cozyreq.cli:main"

[dependency-groups]
dev = [
    "basedpyright>=1.36.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-textual-snapshot>=1.0.0",
    "ruff>=0.14.10",
    "textual-dev>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.basedpyright]
include = ["src", "tests"]
extraPaths = ["src"]

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]

[tool.uv]
package = true
